requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
albucore==0.0.23
albumentations==1.0.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
backoff==2.2.1
banks==2.1.1
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.0
boto3==1.37.28
botocore==1.37.28
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.5.0
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
docker==7.1.0
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-genai==1.9.0
greenlet==3.1.1
griffe==1.7.2
grpcio==1.67.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
identify==2.6.9
idna==3.10
imageio==2.37.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-experimental==0.3.4
langchain-milvus==0.1.9
langchain-openai==0.3.7
langchain-pinecone==0.2.3
langchain-tests==0.3.12
langchain-text-splitters==0.3.6
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.3.11
language_data==1.3.0
lazy_loader==0.4
Levenshtein==0.27.1
lightning==2.5.1
lightning-utilities==0.14.2
llama-cloud==0.1.17
llama-cloud-services==0.6.9
llama-index-core==0.12.28
llama-index-readers-file==0.4.7
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
markdownify==0.13.1
marker-pdf==1.6.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.12
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
munch==4.0.0
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
nougat-ocr==0.1.17
numpy==1.26.4
olefile==0.47
onnx==1.17.0
onnxruntime==1.21.0
openai==1.65.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pdf2markdown4llm==0.1.9
pdfminer.six==20231228
pdfplumber==0.11.5
pdftext==0.6.2
pexpect==4.9.0
pi_heif==0.22.0
pillow==10.4.0
pinecone==5.4.2
pinecone-client==6.0.0
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyheif==0.8.0
pymilvus==2.5.6
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdf==5.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-Levenshtein==0.27.1
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytorch-lightning==2.5.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sconf==0.2.5
sentencepiece==0.2.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
stack-data==0.6.3
streamlit==1.42.2
stringzilla==3.12.3
striprtf==0.0.26
surya-ocr==0.13.1
sympy==1.13.1
syrupy==4.8.2
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
timm==0.5.4
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.6.0
torchmetrics==1.7.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
unstructured==0.17.2
unstructured-client==0.32.0
unstructured-inference==0.8.10
unstructured.pytesseract==0.3.15
urllib3==2.3.0
virtualenv==20.30.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
